refactor : LottoGame 클래스로 정리
